Extract Bedrock code & dependencies into a separate Spring autoconfigure module
