+++ embabel-agent-api/src/main/resources/logback-spring.xml
@@ -21,8 +21,6 @@
     <logger name="com.embabel.agent.api.annotation.support.AgentMetadataReader" level="INFO"/>
     <logger name="com.embabel.agent.spi.support.LlmRanker" level="INFO"/>
     <logger name="com.embabel.agent.spi.support.com.embabel.agent.spi.support.springai.ChatClientLlmOperations" level="INFO"/>
-    <logger name="org.springframework.ai.bedrock" level="ERROR"/>
-    <logger name="software.amazon.awssdk" level="ERROR"/>
 
     <!-- 🔍 Root Logger -->
     <root level="INFO">
