+++ embabel-agent-api/src/test/resources/application.properties
@@ -41,5 +41,3 @@ logging.level.com.embabel.agent.core.support.BlackboardWorldStateDeterminer=INFO
 logging.level.com.embabel.agent.api.annotation.support.AgentMetadataReader=INFO
 logging.level.com.embabel.agent.spi.support.LlmRanker=INFO
 logging.level.com.embabel.agent.spi.support.com.embabel.agent.spi.support.springai.ChatClientLlmOperations=INFO
-logging.level.org.springframework.ai.bedrock=ERROR
-logging.level.software.amazon.awssdk=ERROR
