+++ embabel-agent-autoconfigure/embabel-agent-platform-autoconfigure/pom.xml
@@ -3,23 +3,18 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>com.embabel.agent</groupId>
-        <artifactId>embabel-agent-starters</artifactId>
-        <version>0.1.3-SNAPSHOT</version>
+        <artifactId>embabel-agent-autoconfigure</artifactId>
+        <version>0.1.2-SNAPSHOT</version>
     </parent>
-    <artifactId>embabel-agent-starter-bedrock</artifactId>
+    <artifactId>embabel-agent-platform-autoconfigure</artifactId>
     <packaging>jar</packaging>
-    <name>Embabel Agent Bedrock Starter</name>
-    <description>Embabel Agent Bedrock Starter</description>
+    <name>Embabel Agent Autoconfiguration platform</name>
+    <description>Spring Boot Auto-Configuration platform for Embabel Agent API</description>
 
     <dependencies>
         <dependency>
             <groupId>com.embabel.agent</groupId>
-            <artifactId>embabel-agent-starter</artifactId>
-        </dependency>
-
-        <dependency>
-            <groupId>com.embabel.agent</groupId>
-            <artifactId>embabel-agent-bedrock-autoconfigure</artifactId>
+            <artifactId>embabel-agent-api</artifactId>
         </dependency>
     </dependencies>
