+++ embabel-agent-autoconfigure/embabel-agent-platform-autoconfigure/src/main/resources/META-INF/spring.factories
