+++ embabel-agent-autoconfigure/embabel-agent-platform-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.embabel.agent.autoconfigure.platform.AgentPlatformAutoConfiguration
