+++ embabel-agent-autoconfigure/embabel-agent-platform-autoconfigure/src/test/java/com/embabel/agent/autoconfigure/platform/AgentPlatformAutoConfigurationIT.java
