+++ embabel-agent-autoconfigure/embabel-agent-platform-autoconfigure/src/test/java/com/embabel/agent/config/annotation/EnableAgentsAnnotationIT.java
