+++ embabel-agent-autoconfigure/models/embabel-agent-bedrock-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+com.embabel.agent.autoconfigure.models.bedrock.AgentBedrockAutoConfiguration
