+++ embabel-agent-autoconfigure/models/embabel-agent-bedrock-autoconfigure/src/main/resources/application-bedrock.yml
@@ -1,0 +1,93 @@
+spring:
+  autoconfigure:
+    exclude:
+      - org.springframework.ai.model.bedrock.converse.autoconfigure.BedrockConverseProxyChatAutoConfiguration
+      - org.springframework.ai.model.bedrock.cohere.autoconfigure.BedrockCohereEmbeddingAutoConfiguration
+      - org.springframework.ai.model.bedrock.titan.autoconfigure.BedrockTitanEmbeddingAutoConfiguration
+  ai:
+    bedrock:
+      aws:
+        region: ${AWS_REGION:}
+        access-key: ${AWS_ACCESS_KEY_ID:}
+        secret-key: ${AWS_SECRET_ACCESS_KEY:}
+        session-token: ${AWS_SESSION_TOKEN:}
+
+embabel:
+  models:
+    bedrock:
+      models:
+        # US region
+        - name: us.anthropic.claude-3-5-sonnet-20240620-v1:0
+          knowledge-cutoff: 2024-04-01
+          input-price: 3.0
+          output-price: 15.0
+        - name: us.anthropic.claude-3-5-sonnet-20241022-v2:0
+          knowledge-cutoff: 2024-07-01
+          input-price: 3.0
+          output-price: 15.0
+        - name: us.anthropic.claude-3-5-haiku-20241022-v1:0
+          knowledge-cutoff: 2024-07-01
+          input-price: 0.8
+          output-price: 4.0
+        - name: us.anthropic.claude-3-7-sonnet-20250219-v1:0
+          knowledge-cutoff: 2024-10-31
+          input-price: 3.0
+          output-price: 15.0
+        - name: us.anthropic.claude-sonnet-4-20250514-v1:0
+          knowledge-cutoff: 2025-03-01
+          input-price: 3.0
+          output-price: 15.0
+        - name: us.anthropic.claude-opus-4-20250514-v1:0
+          knowledge-cutoff: 2025-03-01
+          input-price: 15.0
+          output-price: 75.0
+        # EU region
+        - name: eu.anthropic.claude-3-5-sonnet-20240620-v1:0
+          knowledge-cutoff: 2024-04-01
+          input-price: 3.0
+          output-price: 15.0
+        - name: eu.anthropic.claude-3-5-sonnet-20241022-v2:0
+          knowledge-cutoff: 2024-07-01
+          input-price: 3.0
+          output-price: 15.0
+        - name: eu.anthropic.claude-3-5-haiku-20241022-v1:0
+          knowledge-cutoff: 2024-07-01
+          input-price: 0.8
+          output-price: 4.0
+        - name: eu.anthropic.claude-3-7-sonnet-20250219-v1:0
+          knowledge-cutoff: 2024-10-31
+          input-price: 3.0
+          output-price: 15.0
+        - name: eu.anthropic.claude-sonnet-4-20250514-v1:0
+          knowledge-cutoff: 2025-03-01
+          input-price: 3.0
+          output-price: 15.0
+        - name: eu.anthropic.claude-opus-4-20250514-v1:0
+          knowledge-cutoff: 2025-03-01
+          input-price: 15.0
+          output-price: 75.0
+        # APAC region
+        - name: apac.anthropic.claude-3-5-sonnet-20240620-v1:0
+          knowledge-cutoff: 2024-04-01
+          input-price: 3.0
+          output-price: 15.0
+        - name: apac.anthropic.claude-3-5-sonnet-20241022-v2:0
+          knowledge-cutoff: 2024-07-01
+          input-price: 3.0
+          output-price: 15.0
+        - name: apac.anthropic.claude-3-5-haiku-20241022-v1:0
+          knowledge-cutoff: 2024-07-01
+          input-price: 0.8
+          output-price: 4.0
+        - name: apac.anthropic.claude-3-7-sonnet-20250219-v1:0
+          knowledge-cutoff: 2024-10-31
+          input-price: 3.0
+          output-price: 15.0
+        - name: apac.anthropic.claude-sonnet-4-20250514-v1:0
+          knowledge-cutoff: 2025-03-01
+          input-price: 3.0
+          output-price: 15.0
+        - name: apac.anthropic.claude-opus-4-20250514-v1:0
+          knowledge-cutoff: 2025-03-01
+          input-price: 15.0
+          output-price: 75.0
