+++ embabel-agent-autoconfigure/models/embabel-agent-bedrock-autoconfigure/src/test/kotlin/com/embabel/agent/config/models/bedrock/BedrockOptionsConverterTest.kt
@@ -1,0 +1,26 @@
+/*
+ * Copyright 2024-2025 Embabel Software, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.embabel.agent.config.models.bedrock
+
+//import com.embabel.agent.config.models.OptionsConverterTestSupport
+//import org.springframework.ai.model.tool.ToolCallingChatOptions
+//// TODO: need a place to access OptionsConverterTestSupport from api and autoconfiguration
+//class BedrockOptionsConverterTest : OptionsConverterTestSupport<ToolCallingChatOptions>(
+//    optionsConverter = BedrockOptionsConverter
+//) {
+//
+//
+//}
