+++ embabel-agent-autoconfigure/pom.xml
@@ -7,39 +7,16 @@
         <version>0.1.3-SNAPSHOT</version>
     </parent>
     <artifactId>embabel-agent-autoconfigure</artifactId>
-    <packaging>jar</packaging>
-    <name>Embabel Agent Autoconfiguration</name>
+    <packaging>pom</packaging>
+    <name>Embabel Agent Autoconfigurations</name>
     <description>Spring Boot Auto-Configurations for Embabel Agent API</description>
     <properties>
         <maven.compiler.source>21</maven.compiler.source>
         <maven.compiler.target>21</maven.compiler.target>
     </properties>
 
-    <dependencies>
-        <dependency>
-            <groupId>com.embabel.agent</groupId>
-            <artifactId>embabel-agent-api</artifactId>
-        </dependency>
-
-        <!--  overriding optional parameter for bedrock -->
-        <dependency>
-            <groupId>org.springframework.ai</groupId>
-            <artifactId>spring-ai-starter-model-bedrock</artifactId>
-            <optional>false</optional>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.ai</groupId>
-            <artifactId>spring-ai-starter-model-bedrock-converse</artifactId>
-            <optional>false</optional>
-        </dependency>
-
-        <!-- Move this one to a dedicated starter once ready-->
-        <dependency>
-            <groupId>com.embabel.agent</groupId>
-            <artifactId>embabel-agent-a2a</artifactId>
-        </dependency>
-
-    </dependencies>
-
+    <modules>
+        <module>embabel-agent-platform-autoconfigure</module>
+        <module>models/embabel-agent-bedrock-autoconfigure</module>
+    </modules>
 </project>
