--- embabel-agent-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,2 +1,0 @@
-com.embabel.agent.autoconfigure.platform.AgentPlatformAutoConfiguration
-com.embabel.agent.autoconfigure.models.AgentBedrockAutoConfiguration
