--- embabel-agent-autoconfigure/src/main/resources/application.yml
@@ -1,6 +1,0 @@
-spring:
-  autoconfigure:
-    exclude:
-      - org.springframework.ai.model.bedrock.converse.autoconfigure.BedrockConverseProxyChatAutoConfiguration
-      - org.springframework.ai.model.bedrock.cohere.autoconfigure.BedrockCohereEmbeddingAutoConfiguration
-      - org.springframework.ai.model.bedrock.titan.autoconfigure.BedrockTitanEmbeddingAutoConfiguration
